Add messaging layer settings to configuration section
