+++ streampipes-pipeline-management/src/main/java/org/apache/streampipes/manager/setup/SpCoreConfigurationStep.java
@@ -27,6 +27,7 @@ public class SpCoreConfigurationStep extends InstallationStep {
     var coreCfg = new DefaultSpCoreConfiguration().make();
 
     StorageDispatcher.INSTANCE.getNoSqlStore().getSpCoreConfigurationStorage().createElement(coreCfg);
+    new StreamPipesEnvChecker().updateEnvironmentVariables();
   }
 
   @Override
