+++ ui/src/app/configuration/messaging-configuration/broker-config/broker-config.component.html
@@ -1,0 +1,35 @@
+<!--
+  ~ Licensed to the Apache Software Foundation (ASF) under one or more
+  ~ contributor license agreements.  See the NOTICE file distributed with
+  ~ this work for additional information regarding copyright ownership.
+  ~ The ASF licenses this file to You under the Apache License, Version 2.0
+  ~ (the "License"); you may not use this file except in compliance with
+  ~ the License.  You may obtain a copy of the License at
+  ~
+  ~    http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  -->
+
+<div fxFlex="100" fxLayout="column">
+    <div class="subsection-title">{{ title }}</div>
+    <div fxLayout="row" fxLayoutGap="15px">
+        <div>
+            <mat-form-field class="form-field" fxFlex="100" color="accent">
+                <mat-label>Host</mat-label>
+                <input matInput [(ngModel)]="host" type="text" />
+            </mat-form-field>
+        </div>
+        <div>
+            <mat-form-field class="form-field" fxFlex="100" color="accent">
+                <mat-label>Port</mat-label>
+                <input matInput [(ngModel)]="port" type="number" />
+            </mat-form-field>
+        </div>
+    </div>
+</div>
