+++ ui/src/app/configuration/messaging-configuration/messaging-configuration.component.ts
@@ -18,11 +18,11 @@
 
 import { Component, OnInit } from '@angular/core';
 import { ConfigurationService } from '../shared/configuration.service';
-import { MessagingSettings } from '../shared/messaging-settings.model';
 import { CdkDragDrop, moveItemInArray } from '@angular/cdk/drag-drop';
 import { SpConfigurationTabs } from '../configuration-tabs';
 import { SpBreadcrumbService } from '@streampipes/shared-ui';
 import { SpConfigurationRoutes } from '../configuration.routes';
+import { MessagingSettings } from '@streampipes/platform-services';
 
 @Component({
     selector: 'sp-messaging-configuration',
