+++ ui/src/app/configuration/shared/configuration.service.ts
@@ -20,12 +20,12 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { MessagingSettings } from './messaging-settings.model';
 import { MultipartUtils } from './multipart-utils';
 import {
+    MessagingSettings,
     SpServiceConfiguration,
     SpServiceRegistration,
-} from '../../../../projects/streampipes/platform-services/src/lib/model/gen/streampipes-model';
+} from '@streampipes/platform-services';
 
 @Injectable()
 export class ConfigurationService {
@@ -51,13 +51,11 @@ export class ConfigurationService {
     }
 
     getMessagingSettings(): Observable<MessagingSettings> {
-        return this.http
-            .get(this.getServerUrl() + '/api/v2/consul/messaging')
-            .pipe(
-                map(response => {
-                    return response as MessagingSettings;
-                }),
-            );
+        return this.http.get(this.getServerUrl() + '/api/v2/messaging').pipe(
+            map(response => {
+                return response as MessagingSettings;
+            }),
+        );
     }
 
     getRegisteredExtensionsServices(): Observable<SpServiceRegistration[]> {
@@ -97,7 +95,7 @@ export class ConfigurationService {
         messagingSettings: MessagingSettings,
     ): Observable<Object> {
         return this.http.post(
-            this.getServerUrl() + '/api/v2/consul/messaging',
+            this.getServerUrl() + '/api/v2/messaging',
             messagingSettings,
         );
     }
